fix(test): switch the path for unit testing to an absolute one
Unit testing works in a containerised environment because the relative path for the workdir is set by the environment to `/tmp`.

Unit testing is currently broken in the IDE because the relative path for the workdir is the basedir of the file you are testing.

The issue is that the code is developed assuming that the workdir is fixed, but during local testing this assumption is violated.