feat(provider): support for the HTTP-01 challenge (#29)
The functionality is designed to activate by setting "webroot" in the provider's configuration. It is the responsibility of system administrators to route HTTP traffic for the prefix /.well-known/acme-challenge/ to Inca.

TODO: There are no unit tests for the providers, and this is partially correct as we do not want to rewrite tests that cover the features provided by the library we are using (lego). However, it would be very helpful to start introducing tests for the glue code.

For now, the introduced functionality is trivial, and it seems possible to safely introduce it without implementing unit testing beforehand for this portion of the code.

NB: Static serving is heavily dependent on the WORKDIR /tmp Docker directive; everything breaks down in different environments.